+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming Next Release
 
+* Fix @exclude on nested types (Michalis)
 * Accept identifiers with Unicode characters (not only ASCII letters) in them, both in UTF-8 and other (like Windows 1252) encodings, following Delphi. See also [announcement](https://github.com/pasdoc/pasdoc/discussions/235) (Michalis)
 * GraphViz classes graph shows implemented interfaces as dashed lines (Michalis)
 * SimpleXML output contains now deprecated / platform / etc. information, and proper unit name (Michalis)